+++ hello_nexus/handler/service_handler.py
@@ -10,9 +10,9 @@ from nexusrpc.handler import StartOperationContext, service_handler, sync_operat
 from temporalio import nexus
 from temporalio.nexus import WorkflowRunOperationContext, workflow_run_operation
 
-from hello_nexus.basic.handler.db_client import MyDBClient
-from hello_nexus.basic.handler.workflows import WorkflowStartedByNexusOperation
-from hello_nexus.basic.service import MyInput, MyNexusService, MyOutput
+from hello_nexus.handler.db_client import MyDBClient
+from hello_nexus.handler.workflows import WorkflowStartedByNexusOperation
+from hello_nexus.service import MyInput, MyNexusService, MyOutput
 
 
 @service_handler(service=MyNexusService)
