+++ pyproject.toml
@@ -106,9 +106,9 @@ packages = [
     "encryption",
     "gevent_async",
     "hello",
-    "hello_nexus",
     "langchain",
     "message_passing",
+    "nexus",
     "open_telemetry",
     "patching",
     "polling",
