+++ tests/hello_nexus/hello_nexus_test.py
@@ -2,31 +2,31 @@ import asyncio
 
 from temporalio.client import Client
 
-import hello_nexus.basic.caller.app
-import hello_nexus.basic.caller.workflows
-import hello_nexus.basic.handler.worker
+import hello_nexus.caller.app
+import hello_nexus.caller.workflows
+import hello_nexus.handler.worker
 from tests.hello_nexus.helpers import create_nexus_endpoint, delete_nexus_endpoint
 
 
 async def test_nexus_service_basic(client: Client):
     create_response = await create_nexus_endpoint(
-        name=hello_nexus.basic.caller.workflows.NEXUS_ENDPOINT,
-        task_queue=hello_nexus.basic.handler.worker.TASK_QUEUE,
+        name=hello_nexus.caller.workflows.NEXUS_ENDPOINT,
+        task_queue=hello_nexus.handler.worker.TASK_QUEUE,
         client=client,
     )
     try:
         handler_worker_task = asyncio.create_task(
-            hello_nexus.basic.handler.worker.main(
+            hello_nexus.handler.worker.main(
                 client,
             )
         )
         await asyncio.sleep(1)
-        results = await hello_nexus.basic.caller.app.execute_caller_workflow(
+        results = await hello_nexus.caller.app.execute_caller_workflow(
             client,
         )
-        hello_nexus.basic.handler.worker.interrupt_event.set()
+        hello_nexus.handler.worker.interrupt_event.set()
         await handler_worker_task
-        hello_nexus.basic.handler.worker.interrupt_event.clear()
+        hello_nexus.handler.worker.interrupt_event.clear()
         print("\n\n")
         print([r.message for r in results])
         print("\n\n")
